backend/requirements.txt
fastapi
uvicorn
openai
pydantic
python-multipart
aiogram
apscheduler
jinja2
python-dotenv
